ate 40 c

+++ Lista_Geral_C/comandos de entrada e saida/question9.c
@@ -4,7 +4,7 @@
 
 int main(){
 
-    int distancia, tempo;
+    float distancia, tempo;
     float velocidade_media;
 
     printf("Informe a distancia percorrida: ");
